test: comments 조회 관련 오류 공유
